+++ tienda/js/main.js
@@ -8,7 +8,7 @@ numeroInput.addEventListener("input", calcularTotal);
 
 function calcularTotal() {
   // Obtener el valor del número ingresado
-  const numero = parseInt(numeroInput.value);
+  const numero = parseFloat(numeroInput.value);
   const precio = parseFloat(_precio.value);
     if (numero < 0 ) {
         totalElement.textContent = 'No se puede calcular, ingrese un valor positivo';
